Partial implementation for https://github.com/zdila/gpx-animator/issues/77
Implement a map tile cache to avoid downloading the same files each time
a slightly different rendering is tried. User specifies location of cache
and lifetime of cached tiles. If no cache is specified then revert to old
behavior of downloading tiles each time renderer is run.

This is a partial implementation because only the CLI has been updated to
allow specification of cache parameters.

Signed-off-by: Tod Fitch <tod@fitchfamily.org>