Adds support for moving map
Moving map is implemented by defining a rectangular viewport over the
existing video. Think of a viewport as a rectangular hole that masks out
most of the video except for the part you can see. The viewport
automatically tries to keep the tracking dot in the middle, ensuring it
never crosses any edges of the original video.

This feature is controlled by --viewport-width and --viewport-height.

Test: java -jar GpXAnimator.jar --input 2020-08-16.gpx
    --tms-url-template 'http://mt1.google.com/vt/lyrs=m&x={x}&y={y}&z={zoom}'
    --background-map-visibility 1.0
    --viewport-height 640 --viewport-width 640
    --total-time 120000 --tail-duration 10000
    --attribution '' --information-location 'bottom left'
    --output movie.mp4